Implement of q_reverse to reverse whole queue
q_reverse is a function used to reverse whole queue and implemented in a
naive way -- simply exchange each node's next and prev will do the
trick.

The above process is done with help of list_for_each_safe, which is a
simple MACRO allow us to iterate over a given list.

The last thing to keep in mind is that don't forget to reverse head's
next and prev, too. As list_for_each_safe won't access head of the list