Implement function to delete duplicated string
d_delete_dup() is a function to delete duplicated string in queue in
following manners:
first let ele1, ele2 become first and second element in queue. Then, try
to compare ele1's and ele2's string to see if they have same str. If
so, use another loop to do:
    delete ele2
    update ele2 to be the next element
    keep deleted ele2 until ele1 and ele2 doesn't have same str.
Break the loop and delete ele1. Update ele1 to be last ele2 which cause
the break of loop(whose str isn't the same with ele1's)

If it turns out ele1 and ele2 doesn't have same str, update ele1 to be
ele2; ele2 to be next ele followed by ele2.

Mind that delete in here also explicity free the memory allocated for
string and nodes.