Implement q_ascend and q_descend
q_descend is a function to Remove every node which has a node with a
strictly greater value anywhere to the right side of it. The alg is
basically:
Iteration from last node to the head.
Save the max string val max_str during the iteration.
If there exist a node whose str value < max_str, means there is a
strictly greater value in the right side of the node. Remove that node.

likewise, q_ascend is another function using same alg as q_descend while
it now make sure to Remove every node which has a node with a strictly
less value anywhere to the right side of it.

Mind that remove in here will also free the allocated mem for str and
element_t var.